refactor(expect-utils, jest-mock): keep FunctionParameters in jest-mock
Self-review pointed out that @jest/expect-utils is a public package
(community matcher libraries import from it), and FunctionParameters is
a 90-line, 15-overload utility whose only consumer is one method on
MockInstance. Hoisting it into expect-utils saddled every downstream
with a permanent public type.

AsymmetricMatcher is genuinely shared protocol infra (the equals()
runtime consumes the protocol) and stays in expect-utils. Move
FunctionParameters and its private helpers (WithAsymmetricMatchers,
DeepAsymmetricMatcher, FunctionParametersInternal) to jest-mock.
expect's MockParameters now imports FunctionParameters from jest-mock,
which is fine since expect already depends on jest-mock at runtime.
