+++ packages/expect-utils/src/index.ts
@@ -10,9 +10,4 @@ export {equals, isA} from './jasmineUtils';
 export type {EqualsFunction} from './jasmineUtils';
 export * from './utils';
 
-export type {
-  AsymmetricMatcher,
-  FunctionParameters,
-  Tester,
-  TesterContext,
-} from './types';
+export type {AsymmetricMatcher, Tester, TesterContext} from './types';
