+++ packages/expect/src/types.ts
@@ -9,11 +9,10 @@
 import type {
   AsymmetricMatcher,
   EqualsFunction,
-  FunctionParameters,
   Tester,
 } from '@jest/expect-utils';
 import type * as jestMatcherUtils from 'jest-matcher-utils';
-import type {MockInstance} from 'jest-mock';
+import type {FunctionParameters, MockInstance} from 'jest-mock';
 import type {INTERNAL_MATCHER_FLAG} from './jestMatchersObject';
 
 export type SyncExpectationResult = {
